feat: :sparkles: step 2개로 job 구성
- 파일 위치 변경
- 2가지의 step으로 job을 구성하도록 함
