--- src/test/java/io/springbatch/springbatch/HelloJobConfiguration.java
@@ -1,36 +1,0 @@
-package io.springbatch.springbatch;
-
-import org.springframework.batch.core.Job;
-import org.springframework.batch.core.Step;
-import org.springframework.batch.core.configuration.annotation.JobBuilderFactory;
-import org.springframework.batch.core.configuration.annotation.StepBuilderFactory;
-import org.springframework.batch.repeat.RepeatStatus;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-@Configuration
-public class HelloJobConfiguration {
-
-    private final JobBuilderFactory jobBuilderFactory;
-    private final StepBuilderFactory stepBuilderFactory;
-
-    @Bean
-    public Job helloJob() {
-        return jobBuilderFactory.get("helloJob")
-                                .start(helloStep())
-                                .build();
-    }
-
-    @Bean
-    public Step helloStep() {
-        return stepBuilderFactory.get("helloStep")
-                                 .tasklet(((contribution, chunkContext) -> {
-                                     System.out.println("Hello Spring Batch");
-                                     return RepeatStatus.FINISHED;
-                                 }))
-                                 .build();
-    }
-}
